Add no_std support
The only behavior changed by disabling the `std` feature is removal of
the `Date::today` and `DateTime::now` functions, since they depend on
`std::time::SystemTime`. All other changes are switching from aliases
in `std` to the underlying item in `core` or `alloc`.

Building without the `std` feature requires rustc 1.89.0 or newer, as
some floating-point math functions were only recently added to `core`
(and are still not available as methods, motivating the trait wrapper).
It also currently requires the `alloc` crate in a few places.